// pronti-frontend-main/front/src/app/modulos/rubros-rol/form/form-rubros-rol.enum.ts
import {CodigoAuxiliarRubrosRolEnum} from '../../../enums/codigo-auxiliar-rubros-rol.enum';
import {UnidadesEnum} from '../../../enums/unidades.enum';
import {SiNoEnum} from '../../../enums/si-no.enum';

export enum FormRubrosRolEnum {
  codigoAuxiliar = 'codigoAuxiliar',
  nombre = 'nombre',
  nombreAuxiliarUno = 'nombreAuxiliarUno',
  nombreAuxiliarDos = 'nombreAuxiliarDos',
  unidad = 'unidad',
  calculaIess = 'calculaIess',
  calculaRenta = 'calculaRenta',
  calculaDecTercero = 'calculaDecTercero',
  calculaDecCuarto = 'calculaDecCuarto',
  calculaFReserva = 'calculaFReserva',
  calculaVacaciones = 'calculaVacaciones',
  seSuma = 'seSuma',
}